makes guest API more similar to host
This makes the parameters to filter function more like those in the host
by creating a proxy interface. Doing so only slows down the no-op case
by some tens of nanoseconds, so it doesn't matter.

The interesting part is if users return nil for status instead of
multiple values it is faster. So, this changes signature to return a
pointer to status. (I checked against stack based value and it is still
faster).

Finally, this optimizes the example, because our test data doesn't
actually have podSpecNodeName field. This drop in execution time proves
parameters not read are lazy.

Signed-off-by: Adrian Cole <adrian@tetrate.io>