Switches from UID to pointer for identifying cycle IDs
Before, we used the UID to identify which cycle is in progress. However,
when a scheduling cycle fails, the same UID can happen the next time.

This uses the pointer instead as this is different per run, and easier
to compare in wasm. We use the lower bits as that's likely to not
conflict, easier to use in wasm, and also can't identify fully the real
host memory address.

Finally, this resets the pointer between benchmark runs. This makes sure
a run isn't accidentally faster. Before, it was faster due to colliding
on the same UID. By resetting the pointer, we avoid re-creating that
issue.

Signed-off-by: Adrian Cole <adrian@tetrate.io>