Add api.Plugin as a base type for guest plugins
This adds `api.Plugin` as a base type for guest plugins. This is similar
to `framework.Plugin`, but doesn't yet define `Name()` as there's no
purpose for it. Notes on that are added to RATIONALE.md.

This mostly gets the guest SDK looking much more like the framework one.
This also makes it easier to test for misconfiguration, such as
registering multiple instances of plugins.

Signed-off-by: Adrian Cole <adrian@tetrate.io>