Run codec benchmarks in wasm
This changes the guest codec benchmarks to actually run in wasm. By
using TinyGo with wazero, we can see performance of unmarshal
realistically.

Note: This doesn't solve for real data yet because I can't get the yaml
to proto converter to work in TinyGo. We may want to check in serialized
protos as they aren't that big.

Signed-off-by: Adrian Cole <adrian@tetrate.io>