config: add support to file:// and http(s):// URIs
Extends the current plugin config to use instead a URI. In the case
of `file://` the behavior is the same as it is currently.
In the case of `http(s)://` it will fetch the URI and try to
evaluate it as a wasm payload.

This PR is based on earlier work on `dapr/component-contrib`.
See: https://github.com/dapr/components-contrib/pull/3005

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>