guest: add plugin.Set for simpler example and fixes URL parsing
This adds `plugin.Set` which makes it easier to write plugins. This also
makes the example that uses it simpler by avoiding nottinygo. This is
simpler because it reduces the amount of flags used, as well a flat
package is easier to write. However, this can't be tested with tinygo
anymore, and is slower.

Hence, this adds an "advanced" example with the former logic, which is
more verbose and tricky, but faster.

This also fixes some bugs in URL parsing due to some logic not copied
over from dapr. Notably, the original logic handled relative paths and
better error messages. This also simplifies the logic further as we were
parsing into URL type, just to convert it back to a string!

Signed-off-by: Adrian Cole <adrian@tetrate.io>